requirements.txt
# Database
sqlalchemy>=2.0.0
asyncmy>=0.2.9
redis>=5.0.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# FastAPI 核心依赖
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# 数据验证 (兼容 Python 3.13)
pydantic>=2.10.0
pydantic-settings>=2.7.0

# 开发工具
python-dotenv>=1.0.0

# LLM Council (HTTP Client)
httpx>=0.27.0

# Market Data
yfinance>=0.2.48
apscheduler>=3.10.0

# MCP Client (Bing Search)
mcp>=1.9.0

